demo resources: create dedicated folder to separate from lib code
